methods for mappings

+++ app/controllers/graduation_plan_mappings_controller.rb
@@ -12,29 +12,27 @@ class GraduationPlanMappingsController < ApplicationController
     render json: GraduationPlanMappingSerializer.new(mappings, options)
   end
 
-  def update
-    if mapping_id
-      mapping = GraduationPlanMapping.find_by_id_and_graduation_plan_id @graduation_plan.id, mapping_id
-    else
-      mapping = GraduationPlanMapping.new
-    end
+  def create
+    mapping = GraduationPlanMapping.new
 
-    ca_id = credit_assignment_id
-    gpr_id = graduation_plan_requirement_id
+    mapping.credit_assignment = CreditAssignment.find_by_id_and_user_id(credit_assignment_id, @graduation_plan.user_id)
 
-    credit_assignment_id = relations[:credit_assignment][:data][:id]
-
-    mapping.credit_assignment = CreditAssignment.find_by_user_id_and_id(@graduation_plan.user_id, credit_assignment_id)
     mapping.graduation_plan = @graduation_plan
-    mapping.graduation_plan_requirement = GraduationPlanRequirement.find graduation_plan_requirement_id
+    mapping.graduation_plan_requirement = GraduationPlanRequirement.find_by_id_and_status graduation_plan_requirement_id, 'active'
+
+    if GraduationPlanMapping.find_by_graduation_plan_id_and_credit_assignment_id(@graduation_plan.id, mapping.credit_assignment_id)
+      render json: { message: 'Duplicate mapping' }, status: 422 and return
+    end
 
     mapping.save!
 
     render json: GraduationPlanMappingSerializer.new(mapping)
   end
 
-  def delete
+  def destroy
     mapping = GraduationPlanMapping.find_by_id_and_graduation_plan_id params[:mapping_id], @graduation_plan.id
+    raise ActiveRecord::RecordNotFound.new if mapping.nil?
+
     mapping.destroy!
 
     render nothing: true, status: 204
@@ -42,13 +40,18 @@ class GraduationPlanMappingsController < ApplicationController
 
 private
   def get_graduation_plan
-    @graduation_plan = GraduationPlan.find_by_user_id(params[:student_id]) ||
-      GraduationPlan.create!(user_id: params[:student_id])
+    @graduation_plan = GraduationPlan.find_by_user_id(params[:student_id])
+    
+    return @graduation_plan if @graduation_plan
+    
+    user = User.find params[:student_id]
+    
+    @graduation_plan = GraduationPlan.create!(user: user)
   end
 
   def mapping_id
     params.require(:data)
-      .permit(:id)
+      .dig(:id)
   end
 
   def credit_assignment_id
